Adding Agent-Wait after creation
After creating the machine, it is now started. The machine is triggered
by the Docker Machine driver in regular intervalls to check if the OS is
currently running. This is implemented by pinging the guest agent.

After the VM is up, the driver crashes hard, because the next steps are
not yet implemented.

Next steps are to get the IP via agent and try to log in.