Next step, IP detection
IP-Detection is now implemented via QEMU guest agent and the VM can now
be destroyed via docker-machine.

Next up is to generate an SSH private/public key pair and transfer the
public part to the authenticated_keys on the VM. According to [1],
the username and password are

        docker / tcuser

and we can use a simple go-internal ssh connection to transfer the keys.

[1] https://github.com/boot2docker/boot2docker#ssh-into-vm